convert to float

+++ inc/load_data.h
@@ -1,5 +1,5 @@
 #include "commun.h"
 
 char load_text(void);
-void print_trace(float *buffer, uint32_t size);
+void print_trace(double *buffer, uint32_t size);
 char load_traces(void);
